Print serial number of CM0+ target
By printing the serial number of the target acted upon enables a lot of use cases like statistical analysis and tracability. The serial number is read from the devices memory using SWD.

Successfully tested on samd21 hardware as well as verified using datasheet of the following targets:

* samd09
* samd10
* samd11
* samd20
* samd21
* samc21
* saml21
* saml22
* samr21
* samr30
* samda1

However the targets samr34 as well as samr35 do not seem to support serial number extraction.